# perf: prover full improvements