--- a/prover/crypto/ringsis/gen.go
+++ b/prover/crypto/ringsis/gen.go
@@ -1,5 +1,0 @@
-package ringsis
-
-//go:generate go run ./templates --logTwoBound=16 --modulusDegree=64
-//go:generate go run ./templates --logTwoBound=8 --modulusDegree=64
-//go:generate go run ./templates --logTwoBound=8 --modulusDegree=32