--- a/prover/crypto/vortex/reedsolomon.go
+++ b/prover/crypto/vortex/reedsolomon.go
@@ -3,9 +3,9 @@ package vortex
 import (
 	"fmt"
 
-	"github.com/consensys/linea-monorepo/prover/maths/common/mempool"
+	"github.com/consensys/gnark-crypto/ecc/bls12-377/fr/fft"
 	"github.com/consensys/linea-monorepo/prover/maths/common/smartvectors"
-	"github.com/consensys/linea-monorepo/prover/maths/fft"
+	wfft "github.com/consensys/linea-monorepo/prover/maths/fft"
 	"github.com/consensys/linea-monorepo/prover/maths/field"
 )
 
@@ -15,7 +15,7 @@ import (
 // the Lagrange basis Omega_{n * blow-up} where blow-up corresponds to the
 // inverse-rate of the code. The code is systematic as the original vector is
 // interleaved within the encoded vector.
-func (p *Params) rsEncode(v smartvectors.SmartVector, pool mempool.MemPool) smartvectors.SmartVector {
+func (p *Params) rsEncode(v smartvectors.SmartVector) smartvectors.SmartVector {
 
 	// Short path, v is a constant vector. It's encoding is also a constant vector
 	// with the same value.
@@ -23,29 +23,44 @@ func (p *Params) rsEncode(v smartvectors.SmartVector, pool mempool.MemPool) smar
 		return smartvectors.NewConstant(cons.Val(), p.NumEncodedCols())
 	}
 
-	// Interpret the smart-vectors as a polynomial in Lagrange form
-	// and returns a vector of coefficients.
-	asCoeffs := smartvectors.FFTInverse(v, fft.DIT, true, 0, 0, pool)
-	if pool != nil {
-		defer func() {
-			if pooled, ok := asCoeffs.(*smartvectors.Pooled); ok {
-				pooled.Free(pool)
-			}
-		}()
+	expandedCoeffs := make([]field.Element, p.NumEncodedCols())
+
+	// copy the input
+	v.WriteInSlice(expandedCoeffs[:v.Len()])
+
+	const rho = 2
+	if rho != p.BlowUpFactor {
+		smallDomain := p.Domains[0]
+		largeDomain := p.Domains[1]
+
+		smallDomain.FFTInverse(expandedCoeffs[:v.Len()], fft.DIF, fft.WithNbTasks(1))
+		fft.BitReverse(expandedCoeffs[:v.Len()])
+
+		largeDomain.FFT(expandedCoeffs, fft.DIF, fft.WithNbTasks(1))
+		fft.BitReverse(expandedCoeffs)
+
+		return smartvectors.NewRegular(expandedCoeffs)
 	}
 
-	// Pad the coefficients
-	expandedCoeffs := make([]field.Element, p.NumEncodedCols())
-	asCoeffs.WriteInSlice(expandedCoeffs[:asCoeffs.Len()])
+	// fast path; we avoid the bit reverse operations and work on the smaller domain.
+	inputCoeffs := field.Vector(expandedCoeffs[:p.NbColumns])
 
-	// This is not memory that will be recycled easily
-	return smartvectors.FFT(smartvectors.NewRegular(expandedCoeffs), fft.DIT, true, 0, 0, nil)
+	p.Domains[0].FFTInverse(inputCoeffs, fft.DIF, fft.WithNbTasks(1))
+	inputCoeffs.Mul(inputCoeffs, p.CosetTableBitReverse)
+
+	p.Domains[0].FFT(inputCoeffs, fft.DIT, fft.WithNbTasks(1))
+	for j := p.NbColumns - 1; j >= 0; j-- {
+		expandedCoeffs[rho*j+1] = expandedCoeffs[j]
+		expandedCoeffs[rho*j] = v.Get(j)
+	}
+
+	return smartvectors.NewRegular(expandedCoeffs)
 }
 
 // IsCodeword returns nil iff the argument `v` is a correct codeword and an
 // error is returned otherwise.
 func (p *Params) isCodeword(v smartvectors.SmartVector) error {
-	coeffs := smartvectors.FFTInverse(v, fft.DIT, true, 0, 0, nil)
+	coeffs := smartvectors.FFTInverse(v, wfft.DIT, true, 0, 0, nil)
 	for i := p.NbColumns; i < p.NumEncodedCols(); i++ {
 		c := coeffs.Get(i)
 		if !c.IsZero() {