Improves on `beta-3.1-rc1`, tested on `23691973-23691985-getZkProof` . Can yields up to ~40% end-to-end proof generation time improvement.

Added smaller benches too for `TransversalHash` and `CommitWithSIS` : 

```
benchmark                     old ns/op      new ns/op      delta
BenchmarkCommitWithSIS-64     3119849850     2655324881     -14.89%

benchmark                     old allocs     new allocs     delta
BenchmarkCommitWithSIS-64     6394983        792091         -87.61%

benchmark                     old bytes      new bytes      delta
BenchmarkCommitWithSIS-64     4607440872     2311013768     -49.84%
```